pyproject.toml
[project]
name = "langgraph-helper-agent"
version = "0.1.0"
description = "AI LangGraph Helper Agent with offline and online modes (Gemini + LangGraph)"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Your Name" }]
license = { text = "MIT" }
dependencies = [
  "langgraph>=0.2.34,<0.3.0",
  "langchain-core>=0.3.11,<0.4.0",
  "google-generativeai>=0.8.3,<0.9.0",
  "tavily-python>=0.5.0,<0.6.0",
  "duckduckgo-search>=6.2.11,<7.0.0",
  "requests>=2.32.0,<3.0.0",
  "pydantic>=2.7.0,<3.0.0",
  "typing-extensions>=4.9.0,<5.0.0",
  "beautifulsoup4>=4.12.0,<5.0.0",
  "faiss-cpu>=1.8.0,<1.9.0",
  "sentence-transformers>=3.0.0,<4.0.0",
  "python-dotenv>=1.0.0,<2.0.0",
  "mcpdoc>=0.0.9,<0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
  "ruff>=0.6.0,<0.7.0",
  "pre-commit>=3.8.0,<4.0.0",
  "mypy>=1.11.0,<2.0.0",
  "bandit>=1.7.5,<2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agent", "retrievers", "utils"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "W", "C90"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
check_untyped_defs = false
disallow_untyped_defs = false
disallow_incomplete_defs = false
